Added effects, refactored damage system to use effect

+++ src/ecs-survivors/Assets/Code/Gameplay/Features/Effects/EffectComponents.cs
@@ -6,5 +6,5 @@ namespace Code.Gameplay.Features.Effects
     [Game] public class DamageEffect : IComponent { }
     [Game] public class ProducerId : IComponent { public int Value; }
     [Game] public class TargetId : IComponent { public int Value; }
-    [Game] public class EffectValue : IComponent { public int Value; }
+    [Game] public class EffectValue : IComponent { public float Value; }
 }
